requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
torch
transformers
sentence-transformers
pillow
python-multipart
pytesseract